requirements.txt
fastapi
uvicorn
requests
pydantic
openai
python-dotenv